refactor : ProductService 의 validateSeller 메소드 이동
상품 관련 메소드가 아닌 판매자 인증 관련 메소드라서,
목적에 맞게 SellerAuthService 로 이동
